Fix problem with diagram not showing in TF registry page
The diagram was not showing in the Terraform Registry page for this
module [1].  Here the diagram link is changed to an absolute URL in
the git repository.

From the Terraform modules documentation [2]:
```
If you are linking to a file or embedding an image contained in the
repository itself, use a commit-specific absolute URL so the link won't
point to the wrong version of a resource in the future.
```

+ [1] https://registry.terraform.io/modules/sciomedes/ec2-arch-linux-instance/aws/1.0.0
+ [2] https://www.terraform.io/docs/modules/index.html